LayoutTests/fast/dom/HTMLElement/attr-false-string.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<div id="div" contentEditable="false"></div>
<script>
description('When contentEditable attribute is "false" string, element.contentEditable returns "false" and the element is NOT editable.')

shouldBe('document.getElementById("div").getAttribute("contentEditable")','"false"');
shouldBe('document.getElementById("div").contentEditable', '"false"');
shouldBe('document.getElementById("div").isContentEditable', 'false');
shouldBe('window.getComputedStyle(div, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

document.getElementById("div").style.display= 'none';
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
