third_party/blink/web_tests/fast/dom/HTMLElement/attr-false-string-expected.txt

When contentEditable attribute is "false" string, element.contentEditable returns "false" and the element is NOT editable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("div").getAttribute("contentEditable") is "false"
PASS document.getElementById("div").contentEditable is "false"
PASS document.getElementById("div").isContentEditable is false
PASS window.getComputedStyle(div, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS successfullyParsed is true

TEST COMPLETE
